refactor(index): split internals.rs into duty-named layer modules
internals.rs (1118 lines, ~44 methods spanning 9 unrelated duties) was a junk
drawer. Split by duty into sibling modules at the same module depth (so the
pub(super)=pub(in index) cross-module reach is preserved — extracting into a
grandchild module would narrow it):

  fts.rs            FTS index lifecycle
  meta.rs           index_meta + content-revision digest
  git_meta.rs       git history / GitHub status persistence
  graph_index.rs    edge/logical-symbol index lifecycle + coverage
  packages.rs       crate-aware package-root refresh
  parser_failures.rs parse-failure bookkeeping
  staleness.rs      staleness detection + lazy-heal reads
  file_lifecycle.rs file rows, scope deletion, counts
  file_index.rs     (renamed from internals.rs) the file->rows indexing pipeline

Private methods now called across the split were bumped to pub(super). Methods
stay on IndexDatabase, so zero call-site changes. No behavior change.
