+++ crates/rag-rat-core/src/index/parser_failures.rs
@@ -1,0 +1,41 @@
+//! Parser-failure bookkeeping: record a failed parse and report failure counts/paths.
+
+use super::*;
+
+impl IndexDatabase {
+    pub(super) fn insert_parser_failure(
+        &self,
+        path: &Path,
+        language: Language,
+        message: &str,
+    ) -> anyhow::Result<()> {
+        self.storage.connection().execute(
+            "INSERT INTO parser_failures(path, language, message) VALUES (?1, ?2, ?3)",
+            params![path_string(path), language.as_str(), message],
+        )?;
+        Ok(())
+    }
+
+    pub(super) fn parser_failure_count(&self) -> anyhow::Result<u64> {
+        let count = self.storage.connection().query_row(
+            "SELECT COUNT(*) FROM parser_failures",
+            [],
+            |row| row.get::<_, i64>(0),
+        )?;
+        Ok(u64::try_from(count).unwrap_or(0))
+    }
+
+    pub(super) fn parser_failure_paths(&self) -> anyhow::Result<Vec<ParserFailure>> {
+        let mut stmt = self.storage.connection().prepare(
+            "SELECT path, language, message FROM parser_failures ORDER BY path, language, message",
+        )?;
+        let rows = stmt.query_map([], |row| {
+            Ok(ParserFailure { path: row.get(0)?, language: row.get(1)?, message: row.get(2)? })
+        })?;
+        let mut failures = Vec::new();
+        for row in rows {
+            failures.push(row?);
+        }
+        Ok(failures)
+    }
+}
